+++ src/components/sidebar/components/SubMenu.tsx
@@ -1,5 +1,5 @@
 import { MenuItemType } from "../../../constants/data";
-import useTabMenu from "../../../hooks/useTabLayout";
+import useTabMenu from "../../../hooks/useTabMenu";
 import { TabMenuListType } from "../../../states/layout/layoutAtom";
 import Multilevel from "./Multilevel";
